refactor initialization into StateInitializer class
